yocto-autobuilder: Consolidated commit
This commit fixes a few known issues:

[Yocto #2029]
[Yocto #2028]
[Yocto #1850]
[Yocto #2191]

It also adds buildhistory support via the conf/autobuilder.conf file
as well as improves the way creating auto.conf and bblayers.conf
are generated, in that, it allows the passing of different build
providers (yocto/oe) or bsp providers (Intel, fsc).

As defaultenv is also used quite a bit, in preparation for splitting
yoctoABConfig up into yoctoABLib and yoctoABBuildSets, I've
begun passing it into functions.