yoctoABConfig: DEST on non-accessible slave output
The destination of artifacts was determined by looking for the
existance of the path. When master has no access to the destination
this will obviously fail.

The solution here is to pickle the last DEST of a build set.

The one issue this may introduce is during the first day of builds
as master obviously has no knowledge of the slave's directory
structure. Once the first day of builds finish, this should be ok.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>