YoctoABConfig: Adt support
[Yocto #1877, #2800]

This commit completes full adtrepo support.

First, we now have the ability to build an adt-installer with a defined
ADTREPO_URL. The autobuilder will now build the installer pointing to
a specific dev repo (in autobuilder.conf) if ADTREPO_GENERATE_INSTALLER and/or
ADTREPO_GENERATE_DEV_INSTALLER is set to true.

If ADTREPO_DEV_POPULATE is set to true it will populate the repo at the url with:
SDK_VERSION-BRANCH-GITHASH.

This allows us to build out multiple repos and test against them.

In addition this commit also fixes some issues around the addition of meta-yocto-bsp
to bblayers by checking for the LCONF version.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>