yoctoABConfig.py: 1.3 upgrades/additional targets
This commit covers a few additional targets and support for them.

This adds bare oe-core builds to the yocto-autobuilder. As the layering
for oe-core is obviously different, it adds correct bblayers.conf as well
as the correct artifacts publishing, makeCheckout and other needed changes
to support oe-core

This also adds the x32 build as well as fixes support for meta-yocto-bsp changes.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>