yocto-autobuilder: More initial commits.
I've got to cleanup the prereqs but this should suffice. SQLAlchemy versioning
is tricksy. Some versions work. Some don't.

This is the first version that can actually run and get through a build successfully.

Working on some basic restructuring, fixing the buildstep class so it
can actually pull from the config file and instantiate the buildstep.*
classes. We have a basic helloworld buildstep, setDest and CheckoutLayers.

I've also added some ideas on SanityTest buildstep and have sketched out the
general structure of the rest of the BuildSteps we'll need.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>