Builder ordering and general cleanup.
The old buildbot allowed us to name builders a certain way and they would
be ordered in that sort order. New version as well as how I instantiate the builders
does not allow that. So, a small hack to buildbot (only one so far) and a line
in our config file allows us to order the builders however we would like.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>