force-list fixes and prop reference target
Two issues here.

First, force-list and extra props were not being included correctly
within the div tag on the force build page. this made them visible
and not hidden by the toggle.

Second, I wanted a reference target that showed how to add properties
to a build. Calling it webhobbish even though it has nothing to do
with webhob.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>