build-appliance: Fix setting srcrev
The srcrev should, according to sgw, default to AUTOREV. Also,
I want to ensure that I always pick up the buildappsrcrev, even
if for some reason, the name of the prop gets munged by a trigger.

This iterates through getProperties().asDict() and sets the srcrev
to the last occurance of any property with buildappsrcrev in the title.

Not the most elegent way of doing it, and I'm sure there is a better way,
but for now, it's functional.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>