GetDistroVersion: Wrong way of doing it
This is actually a good point to make. When dealing with buildbot
avoid the temptation to do things like pretend the file you want
to do is local. if it's on the builder, you don't have access to
it.

Which means everything needs to be done "the buildbot way" via the
interfaces provided.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>