CreateAutoConf.py: Squish an exception error.
We have to pass command in as a list due to some weird issue where
build progress will throw an assert error in stepFinished. This does
not happen all the time but when it does happen, just break command
up into a list.

This is a long standing issue that I've seen way back in early 0.8.x
buildbot.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>