CheckOutLayers.py: race condition causes random checkout failures
CheckOutLayers was a nasty hack that pulled in most of the buildbot
Git class via copy and paste (it was reallly dirty code).

That obviously caused some issues with a really hard to reproduce
race condition. So, I cleaned this up. I've not seen the race
condition, so, it should, in theory, be fixed, but I won't be certain
until it's in production (since that's where it occurred).

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>