nightly-multilib.conf: add more builds to multilib target
Add a build for core-image-sato (with IMAGE_INSTALL_append = " lib32-connman-gnome").
A qemuimage test in poky (and a RunSanityTests step here) needs to be added at some point,
so that the test is complete.
Also add builds for lib64-core-image-{sato,sato-sdk,lsb-sdk}.
lib64-core-image-sato it's a safe build, but sato-sdk usually fails.

Signed-off-by: Stefan Stanacar <stefanx.stanacar@intel.com>
Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>