buildset-config: Adding minnow targets
These add the minnow targets to the main autobuilder buildsets.

A few notes:

As minnow is only danny for now, these default to the danny
branch. Once there is a master branch, repo should be set to
master.

This is not in the main nightly config yet. Once I get a good build
with them they will be.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>