Implementation of LAYERVERSION_*
[YOCTO #3982]

This is a fairly large commit that speaks to a long standing issue
of breaking changes that cause issues in the autobuilder.

https://bugzilla.yoctoproject.org/show_bug.cgi?id=3982

This makes it so that we can change a LAYERVERSION_* if
a change in a layer causes a backwards compatibility issue.

In this set of commits we address the changes made to sanity
testing which cause backwards compatibility issues as well
as changes to the toolchain builds.

https://bugzilla.yoctoproject.org/show_bug.cgi?id=4782

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>