YoctoMailer.py: Support branch/pokyrepo definitions
Sometimes you just want the mailer to mail on only specific
branchs/repos. This commit adds the functionality by overriding
parts of the BuildBot mailer. This is invoked by two additional
list parameters to the YoctoMailer class:

from YoctoMailer import YoctoMailNotifier
import Yocto_Message_Formatter

c['status'].append(YoctoMailNotifier(fromaddr="autobuilder@localhost",
               sendToInterestedUsers=False,
               yoctorepos=['git://git.yoctoproject.org/poky',
                           'git://git.yoctoproject.org/eclipse-poky',
                           'git://git.openembedded.org/openembedded-core'],
               branch=['master', 'dylan', 'danny'],
               mode='failing',
               extraRecipients=["root@localhost"],
               messageFormatter=Yocto_Message_Formatter.message_formatter))

These are not required. If left unset, the original Mailer class
functionality is returned. You must set yoctorepos however to get
branch functionality (it really makes no sense to limit builds based
on branches without a yoctorepo set).

This commit also fixes the SourceStamp for the build by setting it to
whatever the info is for a poky/oecore/eclipse-poky layer. There
is a better way to do this, but for now, we're going to cheat a bit.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>