grids: fix how we visual properties
As we can do mix and match builds, we can't put build
properties in the sourcestamp area. Besides the fact, that it is
out of context, it becomes wildly problematic.

This fixes some of this. The long term fix is to fix how we
do SourceStamps. Codebases might be the solution here, as well
as a patch to the ss. context.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>