buildset-config/nightly-qa-extras: make it simpler and usable
This target was missing from the nightly target for some reason.
Also, it wasn't up to date with latest changes. I've simplified it a bit,
so it has only two builds of core-image-minimal with features that are known to work.
And these being core-image-minimal the sanity tests are just boot+ping.

Signed-off-by: Stefan Stanacar <stefanx.stanacar@intel.com>
Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>