buildsteps/CheckOutLayers: use a commit object for git rev-parse
When the object passed to git rev-parse is a tag, we need to dereference
the tag until a commit id is found or the object cannot be dereferenced
anymore hence add ^0. (<rev>^0 is a short-hand for <rev>^{commit})

This avoids confusing tarballs like
http://autobuilder.yoctoproject.org/pub/releases/dora-10.0.0/poky-319e24a8b1e100e950b07bbcad5fc81c42396845.tar.bz2
which have the SHA-1 hash of the tag instead of the commit.
It was first mentioned here
https://lists.yoctoproject.org/pipermail/yocto/2013-October/018607.html

Signed-off-by: Stefan Stanacar <stefanx.stanacar@intel.com>