SetDest: Ensure DestDir is set
DestDir was not getting set for triggered builds. This ensures
that it is set.

Also, in CheckOutLayers ^0 is only working for tag checkout
builds. Will need to revisit this to figure out where things
broke, but in the interest of making sure master is functional,
I'm going to pull it out for now.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>