nightly.conf: Change how tiggers are done.
This commit and the next need to be taken together. This drastically
changes how we do Triggered builds.

This is needed in order to remove the YoctoBlocker while still being
able to trigger multiple builds. Essentially, the triggers are
grouped together and triggered at once.

Nightly now has two groups of triggers. waitForFinish = True (which runs
first) and False (which runs while we're doing the package-indexing.

This should actually substantially reduce the time needed for nightly.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>