yoctogit: Fix broken retryFetch logic
If the fetch fails, we try it again and then ignore the return value? I'm
not sure why this would help but regardless, this logic is just plain wrong.

This fixes the code to try again and this time, check the value. If it fails
we might resort to clobbering if enabled.

This stops teh fetcher failing but the build proceeding regardless.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>