Autobuilder.py/Buildset.py: More sane parser/UI
The parser for Autobuilder.py didn't dive very deep on triggered
builds. If it found one, it just grabbed it's properties and
then displayed them on the UI. This allowed us to mix and match
repos/branches/commits.

The utility of mix and match was needed, but at this point, we
don't really need it. So, this commit removes mix and match,
makes Autobuilder's parser recursive, simplifies commit data
and makes it so we can have trigger of triggers and that the
properties are inherited all the way to the top.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>