nightly.conf: Add additional parameters
This allows us to set the prefered kernel via the UI as well
as supports additional release related parameters that are not
currently supported by PublishArtifacts/SetDest.

This also supports TriggerBuilds schedulerNames_nowait. This
allows us to delay the schedulers of certain buildsets a few seconds
before pushing them out to be built.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>