yoctogit.py: fix callbacks and instance attrs
The callbacks in barecopy and movecopy were wrong. The big issue
this fixes is adding a correct resetWorkdir to each of these methods.

This ensures we are in the right place, doing the right things
depending on the name of the layer. For example, we need to be in
build for poky, but build/meta-qt3 for meta-qt3.

Also, this stops us from changing instance attrs.

Setting instance attributes (like self.todir self.workdir) when
you have a bunch of callbacks creates uglyness beyond belief.

Don't use them. Use local vars.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>