yoctogit.py: Fix timeouts.
We should really never need this big of a timeout, but
poky-contrib (despite my continual kvetching) can be a
large clone, or an autobuilder can have a large build
directory and the mv across filesystems *which you should
never do!* can go slow.

So let's set this sufficently high enough that we don't
have to have failures from a timeout, but sufficently low
enough that if things go south we don't wait forever to
know about it.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>