nightly-deb/ipk: don't use the default tests
We can't use RunSanityTests as it is because the default
test suites (which are defined in testimage.bbclass) include rpm and smart.
Anything included in TEST_SUITES is a mandatory test and will be a fail in this
buildset as obviously this images don't have rpm/smart.
Also we can't use 'ping auto' either (which will run all the tests suitable for the image),
because for core-image-sato-sdk that will run cvs/iptables/sudoku build tests which
take a long time (and we have another buildset nightly-qa-targetbuild for that) so we need
to hardcode the list of tests - the same as the defaults for a rpm core-image-sato-sdk,
minus rpm and smart (and without smart this RunSanity will take half of the time)

Signed-off-by: Stefan Stanacar <stefanx.stanacar@intel.com>