master.py/botmaster.py: use builders list for builders order
This patch uses the order of the list builderNames when requesting
builder names. This order is relevant for the Waterfall order.

However, the order of builderNames is not meaningful by default,
hence the change in botmaster.py makes sure the builderNames is
taken form the configuration files list of builder.

Hence we can get rid of the environment work around (which also
did not honor non Autobuilder builders).

Signed-off-by: Stefan Agner <stefan@agner.ch>