yocto-autobuilder: Remove problematic terms from base code
I've taken some preliminary steps to remove some problematic
terms from the autobuilder code base. Instead of refering to
master/slave we're going to be refering to controller/worker.

This is only going to touch the yocto-autobuilder code base.
For projects that we include within the codebase, people within
those communities will need to make their own decisions, however
I feel this is a worthwhile thing to do.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>