PublishArtifacts.py: GEN_IMG_MD5 optional md5sum generation
The way we do md5sum generation isn't working for QA. This adds
GEN_IMG_MD5 to config and makes it so that if it's True it will
generate and publish the md5sums of the image.

Also, I added it to the docs and fixed the horrible single line
html file. Ick.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>