CreateGPLTarball.py: reference step for source archiver.
This is a demo step on how to utilize the oe-core source
archiver. It's specific to dizzy and will not work right
prior to dizzy.

I'm also including a reference conf called
"awesomefirmware", which is just genericx86-64 but has
a custom property which allows the end user to turn tarball
creation on and off.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>